digital_signature_api/BL/requirements.txt
chardet
kafka
pandas
sqlalchemy
mysqlclient
pdfminer.six
